rename: move screen files to lib/screens
