+++ lib/screens/competition_screen.dart
